complete challenge #11: It saves the entry station
It replaces variable station with a double station. It exposes entry_station instance variable using an attribute reader. It makes the card forget the entry station on touch out by setting it to nil. It refactors to remove in_journey variable. Instead, re-introducing the in_journey? method to infer its status based on the presence of an antry station